transform-declaration-production/requirements.txt
osiris-sdk
apache-beam
